iblups/requirements.txt
django==1.11
djangorestframework==3.6.4
django-stdimage==2.4.1
django-storages==1.5.2
gunicorn==19.7.1
pillow==4.1.0
django-cors-headers==2.1.0
django-extensions==1.7.8
python-dateutil==2.6.0
django-autoslug==1.9.3
mysqlclient==1.3.12
PyJWT==1.5.2
orator==0.9.7
requests==2.18.4